Added backend support for signing up and logging in
